Fixed new user create error.
Fixed an error where if there is no user and they are creating their first task, the task post request doesnt wait for the user to first be created. So the database throws an error when you try to create a task with no user. A simple await on the user creation fixed the problem.